Hide delete button from <SummaryStep>
Remove rendering of delete button in SummaryStep's usage of <LegeerklæringFileList />.
Solved by making a separate AttachmentListWithDeletion component and ensuring delete button
rendering in AttachmentListElement is optional (via an optional render function property to
AttachmentListElement).

+++ src/app/components/legeerklæring-file-list/LegeerklæringFileList.tsx
@@ -8,8 +8,10 @@ import { deleteFile } from '../../api/api';
 import { containsAnyUploadedAttachments } from '../../utils/attachmentUtils';
 import Box from '../box/Box';
 import { Normaltekst } from 'nav-frontend-typografi';
+import AttachmentListWithDeletion from '../attachment-list-with-deletion/AttachmentListWithDeletion';
 
 interface LegeerklæringAttachmentListProps {
+    includeDeletionFunctionality: boolean;
     wrapNoAttachmentsInBox?: boolean;
 }
 
@@ -17,7 +19,8 @@ type Props = LegeerklæringAttachmentListProps & ConnectedFormikProps<Field>;
 
 const LegeerklæringAttachmentList: React.FunctionComponent<Props> = ({
     formik: { values, setFieldValue },
-    wrapNoAttachmentsInBox
+    wrapNoAttachmentsInBox,
+    includeDeletionFunctionality
 }) => {
     const legeerklæring: Attachment[] = values[Field.legeerklæring];
 
@@ -29,24 +32,27 @@ const LegeerklæringAttachmentList: React.FunctionComponent<Props> = ({
         return noAttachmentsText;
     }
 
-    return (
-        <AttachmentList
-            attachments={legeerklæring}
-            onRemoveAttachmentClick={(attachment: Attachment) => {
-                attachment.pending = true;
-                setFieldValue(Field.legeerklæring, legeerklæring);
-                deleteFile(attachment.url!).then(
-                    () => {
-                        setFieldValue(Field.legeerklæring, removeElementFromArray(attachment, legeerklæring));
-                    },
-                    () => {
-                        setFieldValue(Field.legeerklæring, removeElementFromArray(attachment, legeerklæring));
-                    }
-                );
-            }}
-            deleteButtonAriaLabel="Fjern vedlegg"
-        />
-    );
+    if (includeDeletionFunctionality) {
+        return (
+            <AttachmentListWithDeletion
+                attachments={legeerklæring}
+                onRemoveAttachmentClick={(attachment: Attachment) => {
+                    attachment.pending = true;
+                    setFieldValue(Field.legeerklæring, legeerklæring);
+                    deleteFile(attachment.url!).then(
+                        () => {
+                            setFieldValue(Field.legeerklæring, removeElementFromArray(attachment, legeerklæring));
+                        },
+                        () => {
+                            setFieldValue(Field.legeerklæring, removeElementFromArray(attachment, legeerklæring));
+                        }
+                    );
+                }}
+            />
+        );
+    } else {
+        return <AttachmentList attachments={legeerklæring} />;
+    }
 };
 
 export default connect<LegeerklæringAttachmentListProps, Field>(LegeerklæringAttachmentList);
